RUBY-2455: Refactoring code to eliminate code duplication
